+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add `created_at` field to response of `/api/v1/subscriptions/invoices` endpoint.
 
+### Changed
+
+- Disable email autolinking in bio and key-value fields.
+
 ## [2.2.0] - 2023-10-22
 
 ### Added